(scc_alloc): %x is unsigned int.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@22905 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/scache.c
@@ -316,7 +316,7 @@ scc_alloc(krb5_context context, const char *name)
 	}
     } else {
 	_krb5_expand_default_cc_name(context, KRB5_SCACHE_DB, &s->file);
-	asprintf(&s->name, "unique-%08x", (unsigned long)s);
+	asprintf(&s->name, "unique-%08x", (unsigned int)s);
     }
     if (s->file == NULL || s->name == NULL) {
 	scc_free(s);
